Modularized the code further
